tests/pyproject.toml
[project]
name = "voyagent-tests"
version = "0.0.0"
description = "Voyagent workspace test suite (dev-only)."
requires-python = ">=3.12"
dependencies = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "sqlalchemy>=2.0",
    "aiosqlite>=0.20",
    "cryptography>=42",
    "svix>=1.39",
    "respx>=0.21",
    "httpx>=0.27",
    "pyjwt[crypto]>=2.9",
    "fastapi>=0.115",
    "voyagent-schemas-canonical",
    "voyagent-schemas-storage",
    "voyagent-agent-runtime",
    "voyagent-api",
    "voyagent-browser-runner",
    "voyagent-driver-vfs",
]

[tool.uv.sources.voyagent-api]
workspace = true

[tool.uv.sources.voyagent-browser-runner]
workspace = true

[tool.uv.sources.voyagent-driver-vfs]
workspace = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# `tests/` is not really a distributable package — it exists so uv can treat
# the directory as a workspace member and resolve dev-time deps. Hatch needs
# at least one package to build a wheel; point it at the already-existing
# `canonical/` subpackage (authored separately).
[tool.hatch.build.targets.wheel]
packages = ["canonical"]

[tool.uv.sources]
voyagent-schemas-canonical = { workspace = true }
voyagent-schemas-storage = { workspace = true }
voyagent-agent-runtime = { workspace = true }
